src/python/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "petab_sciml"
version = "0.0.1"
authors = [
  {name = "Sebastian Persson"},
  {name = "Dilan Pathirana"},
]
dependencies = [
  "mkstd>=0.0.10",
  "torch",
  "pandas",
  "pydantic",
  "python-libsbml",
  "ruamel.yaml",
]
requires-python = ">=3.8"
description = "Specify parameter estimation problems and hybrid models."
readme = "README.md"
license = {file = "LICENSE.txt"}  # TODO
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

[project.optional-dependencies]
doc = [
    "sphinx>=8.1.3",
    "sphinxcontrib-napoleon>=0.7",
    "sphinx-markdown-tables>=0.0.15",
    "sphinx-rtd-theme>=0.5.1",
    "recommonmark>=0.7.1",
    # pin until ubuntu comes with newer pandoc:
    # site-packages/nbsphinx/__init__.py:1058: RuntimeWarning: You are using an unsupported version of pandoc (2.9.2.1).
    # Your version must be at least (2.14.2) but less than (4.0.0).
    "nbsphinx>=0.9.5",
    "pandoc>=2.4",
    "nbconvert>=7.16.4",
    "ipykernel>= 6.23.1",
    "ipython>=7.21.0",
    "readthedocs-sphinx-ext>=2.2.5",
    "sphinx-autodoc-typehints",
]

test = [
    "pytest"
]

[project.urls]
Repository = "https://github.com/sebapersson/petab_sciml"
"Bug Tracker" = "https://github.com/sebapersson/petab_sciml/issues"
